# openapi-spec.yaml
openapi: 3.0.3
info:
  title: P4 B2B Marketplace API
  description: API for the P4 GCC/MENA B2B Marketplace
  version: 1.0.0
  contact:
    name: P4 Development Team
    email: dev@p4-marketplace.com
servers:
  - url: https://b2b-marketplace-dcd9azhpefdkdve4.canadacentral-01.azurewebsites.net
    description: Production server
...
paths:
...
  /api/catalog/products:
    get:
      tags:
        - Catalog
      summary: Browse products
      description: Browse products with optional search, filtering, and pagination
      security:
        - bearerAuth: []
      parameters:
        - name: keyword
          in: query
          required: false
          schema:
            type: string
            example: "wireless headphones"
        - name: vendorId
          in: query
          required: false
          schema:
            type: string
            example: "01F123456789ABCDEFGH"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            example: "name"
        - name: sortOrder
          in: query
          required: false
          schema:
            type: string
            enum: [ASC, DESC]
            default: "ASC"
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
    post:
      tags:
        - Catalog
      summary: Create a product
      description: Create a new product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
...
  /api/catalog/attributes:
    post:
      tags:
        - Catalog
      summary: Create a product attribute
      description: Create a new product attribute
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductAttribute'
      responses:
        '201':
          description: Product attribute created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
    get:
      tags:
        - Catalog
      summary: Get all product attributes
      description: Retrieve a list of all product attributes with pagination
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Product attributes retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /api/catalog/attributes/{id}:
    get:
      tags:
        - Catalog
      summary: Get product attribute by ID
      description: Retrieve a specific product attribute by its ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "01F123456789ABCDEFGH"
      responses:
        '200':
          description: Product attribute retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
        '404':
          description: Product attribute not found
        '500':
          description: Internal server error
    put:
      tags:
        - Catalog
      summary: Update a product attribute
      description: Update an existing product attribute
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "01F123456789ABCDEFGH"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductAttribute'
      responses:
        '200':
          description: Product attribute updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '404':
          description: Product attribute not found
        '500':
          description: Internal server error
    delete:
      tags:
        - Catalog
      summary: Delete a product attribute
      description: Delete an existing product attribute
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "01F123456789ABCDEFGH"
      responses:
        '200':
          description: Product attribute deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
        '404':
          description: Product attribute not found
        '500':
          description: Internal server error
  /api/catalog/media-assets:
    post:
      tags:
        - Catalog
      summary: Create a media asset
      description: Create a new media asset
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaAsset'
      responses:
        '201':
          description: Media asset created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
    get:
      tags:
        - Catalog
      summary: Get all media assets
      description: Retrieve a list of all media assets with pagination
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Media assets retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /api/catalog/media-assets/{id}:
    get:
      tags:
        - Catalog
      summary: Get media asset by ID
      description: Retrieve a specific media asset by its ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "01F123456789ABCDEFGH"
      responses:
        '200':
          description: Media asset retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
        '404':
          description: Media asset not found
        '500':
          description: Internal server error
    put:
      tags:
        - Catalog
      summary: Update a media asset
      description: Update an existing media asset
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "01F123456789ABCDEFGH"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaAsset'
      responses:
        '200':
          description: Media asset updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '404':
          description: Media asset not found
        '500':
          description: Internal server error
    delete:
      tags:
        - Catalog
      summary: Delete a media asset
      description: Delete an existing media asset
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "01F123456789ABCDEFGH"
      responses:
        '200':
          description: Media asset deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
        '404':
          description: Media asset not found
        '500':
          description: Internal server error
...
  /api/catalog/product-media/product/{productId}:
    get:
      tags:
        - Catalog
      summary: Get media assets for a product
      description: Retrieve all media assets associated with a specific product
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            example: "01F123456789ABCDEFGH"
      responses:
        '200':
          description: Media assets retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
        '404':
          description: Product not found
        '500':
          description: Internal server error
  /api/catalog/products/{id}:
    get:
      tags:
        - Catalog
      summary: Get product by ID
      description: Retrieve a specific product by its ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "01F123456789ABCDEFGH"
      responses:
        '200':
          description: Product retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
        '404':
          description: Product not found
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        data:
          type: object
          description: Response data payload
          example: {}
        success:
          type: boolean
          description: Whether the request was successful
          example: true
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the response
          example: "2025-10-19T06:30:09.284038836Z"
        metadata:
          type: object
          description: Additional metadata
          example: {"message": "Operation successful"}
        error:
          type: object
          description: Error details if success is false
          properties:
            type:
              type: string
              example: "about:blank"
            title:
              type: string
              example: "Not Found"
            status:
              type: integer
              example: 404
            detail:
              type: string
              example: "Resource not found"
            instance:
              type: string
              example: "/api/resource/123"
...
    ProductRequest:
      type: object
      properties:
        name:
          type: string
          example: "Wireless Headphones"
        description:
          type: string
          example: "High quality wireless headphones"
        sku:
          type: string
          example: "WH-12345"
        vendorId:
          type: string
          example: "01F123456789ABCDEFGH"
        price:
          $ref: '#/components/schemas/Money'
        stockQuantity:
          type: integer
          example: 100
        minOrderQuantity:
          type: integer
          example: 1
        weight:
          type: number
          format: decimal
          example: 0.5
        dimensionsLength:
          type: number
          format: decimal
          example: 20.0
        dimensionsWidth:
          type: number
          format: decimal
          example: 15.0
        dimensionsHeight:
          type: number
          format: decimal
          example: 8.0
        productStatus:
          type: string
          enum: [DRAFT, ACTIVE, INACTIVE, DISCONTINUED]
          example: "ACTIVE"
        isActive:
          type: boolean
          example: true
    ProductAttribute:
      type: object
      properties:
        id:
          type: string
          example: "01F123456789ABCDEFGH"
        name:
          type: string
          example: "color"
        displayName:
          type: string
          example: "Color"
        description:
          type: string
          example: "The color of the product"
        attributeType:
          type: string
          enum: [TEXT, NUMBER, BOOLEAN, DATE, SELECT]
          example: "SELECT"
        isRequired:
          type: boolean
          example: false
        isFilterable:
          type: boolean
          example: true
        sortOrder:
          type: integer
          example: 1
    MediaAsset:
      type: object
      properties:
        id:
          type: string
          example: "01F123456789ABCDEFGH"
        originalFilename:
          type: string
          example: "product-image.jpg"
        storagePath:
          type: string
          example: "/images/products/12345.jpg"
        contentType:
          type: string
          example: "image/jpeg"
        fileSize:
          type: integer
          example: 102400
        altText:
          type: string
          example: "Product image"
        caption:
          type: string
          example: "High quality product image"
        mediaType:
          type: string
          enum: [IMAGE, VIDEO, DOCUMENT, OTHER]
          example: "IMAGE"
...
    Money:
      type: object
      properties:
        amount:
          type: string
          example: "29.99"
        currency:
          type: string
          example: "USD"
security:
  - bearerAuth: []